+++ TinkoffASDKUI/TinkoffASDKUI/BusinessLogic/TinkoffPay/Controller/ITinkoffPayController.swift
@@ -11,5 +11,6 @@ import TinkoffASDKCore
 protocol ITinkoffPayController: AnyObject {
     var delegate: TinkoffPayControllerDelegate? { get set }
 
-    func performPayment(paymentFlow: PaymentFlow, method: TinkoffPayMethod)
+    @discardableResult
+    func performPayment(paymentFlow: PaymentFlow, method: TinkoffPayMethod) -> Cancellable
 }